perf(webpack): improve static pages size by using file-loader always instead of url-loader
Webpack use url-loader to load small images (<8 kb), but this leads to bigger static pages.

Imagine you have a 6KB icon 10 times on the page.
By using url-loader, it will be 60kB in the HTML file.
By using file-loader, it will be ~1kB in the HTML file + static file to be downloaded.

I do this optimization because Netlify fails to upload ~60k pages in 30 minutes timeout.
By using this optimisation I hope to reduce the page size and Netlfy will succeed.